Implement `TryBinaryOperator` for `Value::U8`
