+++ core/src/data/value/binary_op/integer/mod.rs
@@ -1,2 +1,4 @@
 mod i8;
+mod u8;
+
 mod macros;
